@@ -14,7 +14,7 @@ description: A new Flutter application.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
   flutter:
@@ -43,6 +43,7 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/icons/album.png
+    - assets/images/google_map_ta.png
 
   fonts:
      - family: Poppin
