@@ -3,10 +3,14 @@ import 'package:flutter/material.dart';
 class PlinkdScaffold extends StatelessWidget {
   final Widget appBar;
   final Widget body;
-  final double curveRaduis;
-  const PlinkdScaffold({Key key, @required this.appBar, @required this.body, this.curveRaduis=20})
+  final double curveRadius;
+  const PlinkdScaffold(
+      {Key key,
+      @required this.appBar,
+      @required this.body,
+      this.curveRadius = 20})
       : super(key: key);
- // git config --global user.name "robincruseo"
+  // git config --global user.name "robincruseo"
 
   @override
   Widget build(BuildContext context) {
@@ -14,9 +18,9 @@ class PlinkdScaffold extends StatelessWidget {
       body: Container(
         decoration: BoxDecoration(
             gradient: LinearGradient(colors: [
-              Color(0XFFe46514),
-              Color(0XFFf79836),
-            ])),
+          Color(0XFFe46514),
+          Color(0XFFf79836),
+        ])),
         child: Column(
           children: <Widget>[scaffoldAppBar(), scaffoldBody()],
         ),
@@ -38,12 +42,19 @@ class PlinkdScaffold extends StatelessWidget {
     return Flexible(
       child: Container(
         child: body,
-        padding: EdgeInsets.all(15),
+        //padding: EdgeInsets.all(15),
         decoration: BoxDecoration(
             color: Colors.white,
+            boxShadow: [
+              BoxShadow(
+                  color: Colors.black.withOpacity(0.1),
+                  blurRadius: 5,
+                  spreadRadius: 5,
+                  offset: Offset(0.5, 0.1))
+            ],
             borderRadius: BorderRadius.only(
-              topLeft: Radius.circular(curveRaduis),
-              topRight: Radius.circular(curveRaduis),
+              topLeft: Radius.circular(curveRadius),
+              topRight: Radius.circular(curveRadius),
             )),
       ),
     );
