@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:plinkd_app/screens/dealScreen.dart';
 
+import 'screens/addFriendScreen.dart';
+import 'screens/peopleMayKnowScreen.dart';
 import 'utils/assets.dart';
 
 void main() => runApp(MyApp());
@@ -13,7 +15,7 @@ class MyApp extends StatelessWidget {
       title: 'Flutter Demo',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(primarySwatch: Colors.blue, fontFamily: poppinFont),
-      home: DealScreen(),
+      home: AddFriendScreen(),
     );
   }
 }
